requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
tqdm
openai
voyageai
umap-learn